Combined Test files to reflect class that provides the functionality beign tested
